requirements.txt
# Existing packages
openai==1.54.3
httpx==0.27.2
httpcore==1.0.6
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
duckduckgo-search==4.4.0

# Streamlit for web interface
streamlit==1.29.0